fix(website): Safe TX, refetch signatures when user signs.
Closes #CAN-663
